+++ src/components/index.jsx
@@ -1,11 +1,12 @@
 export * from "./AppLayout";
-export * from "./Nav";
-export * from "./Coin";
-export * from "./Balance";
-export * from "./VendProduct";
-export * from "./VendController";
-export * from "./InsertCoin";
-export * from "./ChangeOutlet";
-export * from "./HistoryBox";
-export * from "./Button";
+export * from "./MainNav";
 export * from "./ToggleDisplay";
+export * from "./common/Button";
+export * from "./common/Nav";
+export * from "./wallet/Coin";
+export * from "./wallet/Balance";
+export * from "./vending-machine/vend-product/VendProduct";
+export * from "./vending-machine/vend-controller/VendController";
+export * from "./vending-machine/vend-controller/InsertCoin";
+export * from "./vending-machine/vend-controller/ChangeOutlet";
+export * from "./vending-machine/vend-controller/HistoryBox";
