+++ src/components/vending-machine/vend-controller/ChangeOutlet.jsx
@@ -1,4 +1,4 @@
-import { Button } from "./Button";
+import { Button } from "components";
 
 function ChangeOutlet() {
   return (
