+++ src/components/vending-machine/vend-controller/HistoryBox.jsx
